+++ R/EmpiricalCalibrationUsingAsymptotics.R
@@ -99,10 +99,12 @@ print.null <- function(x, ...) {
 #'                  estimate>))/qnorm(0.025)
 #' @param null      An object of class \code{null} created using the \code{fitNull} function or an
 #'                  object of class \code{mcmcNull} created using the \code{fitMcmcNull} function.
+#' @param twoSided  Compute two-sided (TRUE) or one-sided (FALSE) p-value?
+#' @param upper     If one-sided: compute p-value for upper (TRUE) or lower (FALSE) bound?
 #' @param ...       Any additional parameters (currently none).
 #'
 #' @return
-#' The two-sided calibrated p-value.
+#' The calibrated p-value.
 #'
 #' @examples
 #' data(sccs)
@@ -116,31 +118,36 @@ print.null <- function(x, ...) {
 #' empirical calibration is needed to correct p-values. Statistics in Medicine 33(2):209-18,2014
 #'
 #' @export
-calibrateP <- function(null, logRr, seLogRr, ...) {
+calibrateP <- function(null, logRr, seLogRr, twoSided = TRUE, upper = TRUE, ...) {
   UseMethod("calibrateP")
 }
 
 
 #' @describeIn
 #' calibrateP Computes the calibrated P-value using asymptotic assumptions.
+#' 
 #' @export
-calibrateP.null <- function(null, logRr, seLogRr, ...) {
+calibrateP.null <- function(null, logRr, seLogRr, twoSided = TRUE, upper = TRUE, ...) {
+  if (length(logRr) != length(seLogRr)) 
+    stop("The logRr and seLogRr arguments must be of equal length")
 
-  oneAdjustedP <- function(logRR, seLogRr, null) {
-    P_upper_bound <- pnorm((null[1] - logRR)/sqrt(null[2]^2 + seLogRr^2))
-    P_lower_bound <- pnorm((logRR - null[1])/sqrt(null[2]^2 + seLogRr^2))
-    2 * min(P_upper_bound, P_lower_bound)
-  }
-
-  adjustedP <- vector(length = length(logRr))
-  for (i in 1:length(logRr)) {
+  calibrateOneP <- function(i) {
     if (is.na(logRr[i]) || is.infinite(logRr[i]) || is.na(seLogRr[i]) || is.infinite(seLogRr[i])) {
-      adjustedP[i] <- NA
+      return(NA)
+    } else
+    pUpperBound <- pnorm((null[1] - logRr[i])/sqrt(null[2]^2 + seLogRr[i]^2))
+    pLowerBound <- pnorm((logRr[i] - null[1])/sqrt(null[2]^2 + seLogRr[i]^2))
+    if (twoSided) {
+      return(2 * min(pUpperBound, pLowerBound))
+    } else if (upper) {
+      return(pUpperBound)
     } else {
-      adjustedP[i] <- oneAdjustedP(logRr[i], seLogRr[i], null)
+      return(pLowerBound)
     }
   }
-  return(adjustedP)
+
+  calibratedP <- sapply(1:length(logRr), calibrateOneP)
+  return(calibratedP)
 }
 
 #' Compute the (traditional) p-value
@@ -153,9 +160,11 @@ calibrateP.null <- function(null, logRr, seLogRr, ...) {
 #' @param seLogRr   The standard error of the log of the effect estimates. Hint: often the standard
 #'                  error = (log(<lower bound 95 percent confidence interval>) - log(<effect
 #'                  estimate>))/qnorm(0.025)
+#' @param twoSided  Compute two-sided (TRUE) or one-sided (FALSE) p-value?
+#' @param upper     If one-sided: compute p-value for upper (TRUE) or lower (FALSE) bound?
 #'
 #' @return
-#' The two-sided (traditional) p-value.
+#' The (traditional) p-value.
 #'
 #' @examples
 #' data(sccs)
@@ -163,7 +172,15 @@ calibrateP.null <- function(null, logRr, seLogRr, ...) {
 #' computeTraditionalP(positive$logRr, positive$seLogRr)
 #'
 #' @export
-computeTraditionalP <- function(logRr, seLogRr) {
+computeTraditionalP <- function(logRr, seLogRr, twoSided = TRUE, upper = TRUE) {
   z <- logRr/seLogRr
-  return(2 * pmin(pnorm(z), 1 - pnorm(z)))  # 2-sided p-value
+  pUpperBound <- 1 - pnorm(z)
+  pLowerBound <- pnorm(z)
+  if (twoSided) {
+    return(2 * min(pUpperBound, pLowerBound))
+  } else if (upper) {
+    return(pUpperBound)
+  } else {
+    return(pLowerBound)
+  }
 }