+++ NEWS.md
@@ -1,10 +1,17 @@
+EmpiricalCalibration 2.2.0
+==========================
+
+Changes
+
+1. Adding option to make p-values one-sided
+
+
 EmpiricalCalibration 2.1.0
 ==========================
 
 Changes
 
 1. Adding computation of expected absolute systematic error. Can be shown in plots.
-
 
 
 EmpiricalCalibration 2.0.2